make the field values widget a functional component
